fix(capacity): allow joins after split - Needed as there will be no `MemberLeft`event when there is a split, and a handful of adults are "lost" as they are promoted.
