+++ src/capacity/rate_limit.rs
@@ -7,10 +7,9 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use super::{Capacity, MAX_CHUNK_SIZE, MAX_SUPPLY};
-use crate::{network::Network, Result};
-use log::{debug, info};
-use sn_data_types::{PublicKey, Token};
-use xor_name::XorName;
+use crate::network::Network;
+use log::debug;
+use sn_data_types::Token;
 
 /// Calculation of rate limit for writes.
 #[derive(Clone)]
@@ -35,27 +34,6 @@ impl RateLimit {
         let all_nodes = self.network.our_adults().await.len() as u8;
 
         RateLimit::rate_limit(bytes, full_nodes, all_nodes, prefix_len)
-    }
-
-    /// Adds this node to the list of full nodes.
-    pub async fn increase_full_node_count(&mut self, node_id: PublicKey) -> Result<()> {
-        self.capacity.increase_full_node_count(node_id).await?;
-        info!("No. of Full Nodes: {:?}", self.full_nodes().await);
-        Ok(())
-    }
-
-    /// Removes a given node from the list of full nodes.
-    pub async fn decrease_full_node_count_if_present(&mut self, node_name: XorName) -> Result<()> {
-        self.capacity
-            .decrease_full_node_count_if_present(node_name)
-            .await?;
-        info!("No. of Full Nodes: {:?}", self.full_nodes().await);
-        Ok(())
-    }
-
-    /// Returns the count of full_nodes.
-    pub async fn full_nodes(&self) -> u8 {
-        self.capacity.full_nodes().await
     }
 
     fn rate_limit(bytes: u64, full_nodes: u8, all_nodes: u8, prefix_len: usize) -> Token {
